backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "siworkgroup-backend"
version = "0.1.0"
description = "SiWorkGroup Backend - OpenClaw Agent Workgroup Platform"
requires-python = ">=3.11"
dependencies = [
    # Web Framework
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "python-multipart>=0.0.9",
    "websockets>=12.0",

    # Database
    "sqlalchemy[asyncio]>=2.0.0",
    "alembic>=1.13.0",
    "asyncpg>=0.29.0",
    "psycopg2-binary>=2.9.9",

    # Redis & Celery
    "redis[hiredis]>=5.0.0",
    "celery[redis]>=5.3.0",

    # Auth & Security
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "pydantic[email]>=2.6.0",
    "pydantic-settings>=2.2.0",

    # Docker SDK
    "docker>=7.0.0",

    # HTTP Client
    "httpx>=0.27.0",

    # Logging
    "structlog>=24.1.0",

    # OpenAI
    "openai>=1.14.0",

    # Utilities
    "python-slugify>=8.0.0",
    "arrow>=1.3.0",
    "orjson>=3.9.0",
]

[project.optional-dependencies]
dev = [
    # Testing
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.27.0",
    "anyio>=4.3.0",
    "factory-boy>=3.3.0",
    "faker>=24.0.0",
    "aiosqlite>=0.20.0",

    # Code Quality
    "ruff>=0.3.0",
    "mypy>=1.9.0",
    "pre-commit>=3.7.0",

    # Type Stubs
    "types-redis>=4.6.0",
    "types-passlib>=1.7.7",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "C4", "UP"]
ignore = ["E501", "B008", "B904"]

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.mypy]
python_version = "3.11"
strict = false
ignore_missing_imports = true
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["app"]
omit = ["tests/*", "alembic/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
]